refactor(phrases,console): refactor the application details form
refactor the application details form, merge advance settings to settings
